# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/diseases": {
      "get": {
        "operationId": "DiseasesController_list",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "To'liq matn qidiruv (FTS)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "icd",
            "required": false,
            "in": "query",
            "description": "ICD-10/11 kod boshlanishi bilan filter",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Kategoriya filter",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Sahifa raqami",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Sahifa hajmi",
            "schema": {
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "audience",
            "required": false,
            "in": "query",
            "description": "Auditoriya filtri",
            "schema": {
              "type": "string",
              "enum": [
                "PATIENT",
                "STUDENT",
                "NURSE",
                "DOCTOR",
                "SPECIALIST",
                "MIXED"
              ]
            }
          },
          {
            "name": "locale",
            "required": false,
            "in": "query",
            "description": "Til (uz/ru/en)",
            "schema": {
              "default": "uz",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Status filtri (admin uchun). Public holatda default PUBLISHED. 'ALL' barchasini ko'rsatadi.",
            "schema": {
              "type": "string",
              "enum": [
                "ALL",
                "DRAFT",
                "REVIEW",
                "APPROVED",
                "PUBLISHED",
                "REJECTED",
                "ARCHIVED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiseaseListResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Feature disabled"
          }
        },
        "summary": "Kasalliklar ro'yxati va qidiruv",
        "tags": [
          "diseases"
        ]
      },
      "post": {
        "operationId": "DiseasesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiseaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiseaseResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Yangi kasallik yaratish",
        "tags": [
          "diseases"
        ]
      }
    },
    "/api/v1/diseases/{slug}": {
      "get": {
        "operationId": "DiseasesController_getOne",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "L1",
                "L2",
                "L3"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiseaseResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Topilmadi yoki feature disabled"
          }
        },
        "summary": "Bitta kasallik (slug bo'yicha)",
        "tags": [
          "diseases"
        ]
      }
    },
...
    "/api/v1/diseases/{id}": {
      "patch": {
        "operationId": "DiseasesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDiseaseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiseaseResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kasallikni yangilash",
        "tags": [
          "diseases"
        ]
      },
      "delete": {
        "operationId": "DiseasesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Kasallikni arxivlash (soft delete)",
        "tags": [
          "diseases"
        ]
      }
    },
...
    "/api/v1/diseases/{slug}/scientists": {
      "get": {
        "operationId": "MetadataController_listScientists",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Disease slug yoki UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Kasallik olimlari ro'yxati",
        "tags": [
          "disease-metadata"
        ]
      }
    },
    "/api/v1/diseases/{id}/scientists": {
      "post": {
        "operationId": "MetadataController_createScientist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Disease UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScientistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Yangi olim qo'shish (EDITOR+)",
        "tags": [
          "disease-metadata"
        ]
      }
    },
    "/api/v1/diseases/{id}/scientists/{scientistId}": {
      "patch": {
        "operationId": "MetadataController_updateScientist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scientistId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScientistDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Olimni yangilash (EDITOR+)",
        "tags": [
          "disease-metadata"
        ]
      },
      "delete": {
        "operationId": "MetadataController_deleteScientist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scientistId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Olimni o'chirish (EDITOR+)",
        "tags": [
          "disease-metadata"
        ]
      }
    },
    "/api/v1/diseases/{slug}/research": {
      "get": {
        "operationId": "MetadataController_listResearch",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Kasallikning tadqiqotlar ro'yxati",
        "tags": [
          "disease-metadata"
        ]
      }
    },
    "/api/v1/diseases/{id}/research": {
      "post": {
        "operationId": "MetadataController_createResearch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResearchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "409": {
            "description": "DOI ziddiyat (kasallik ichida takrorlangan)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Yangi tadqiqot qo'shish (EDITOR+)",
        "tags": [
          "disease-metadata"
        ]
      }
    },
    "/api/v1/diseases/{id}/research/{researchId}": {
      "patch": {
        "operationId": "MetadataController_updateResearch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "researchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResearchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tadqiqotni yangilash (EDITOR+)",
        "tags": [
          "disease-metadata"
        ]
      },
      "delete": {
        "operationId": "MetadataController_deleteResearch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "researchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tadqiqotni o'chirish (EDITOR+)",
        "tags": [
          "disease-metadata"
        ]
      }
    },
    "/api/v1/diseases/{slug}/genetics": {
      "get": {
        "operationId": "MetadataController_listGenetics",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Kasallikning genetik profili",
        "tags": [
          "disease-metadata"
        ]
      }
    },
    "/api/v1/diseases/{id}/genetics": {
      "post": {
        "operationId": "MetadataController_createGenetic",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGeneticDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Yangi genetik yozuv qo'shish (EDITOR+)",
        "tags": [
          "disease-metadata"
        ]
      }
    },
    "/api/v1/diseases/{id}/genetics/{geneticId}": {
      "patch": {
        "operationId": "MetadataController_updateGenetic",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "geneticId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGeneticDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Genetik yozuvni yangilash (EDITOR+)",
        "tags": [
          "disease-metadata"
        ]
      },
      "delete": {
        "operationId": "MetadataController_deleteGenetic",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "geneticId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Genetik yozuvni o'chirish (EDITOR+)",
        "tags": [
          "disease-metadata"
        ]
      }
    },
...
  },
  "info": {
    "title": "MedSmart Pro API",
    "description": "Tibbiy axborot tizimi API hujjatlari",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DiseaseResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "icd10": {
            "type": "string"
          },
          "icd11": {
            "type": "object",
            "nullable": true
          },
          "nameUz": {
            "type": "string"
          },
          "nameRu": {
            "type": "object",
            "nullable": true
          },
          "nameEn": {
            "type": "object",
            "nullable": true
          },
          "nameLat": {
            "type": "object",
            "nullable": true
          },
          "synonyms": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "category": {
            "type": "string"
          },
          "audienceLevels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "severityLevels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "protocolSources": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string"
          },
          "editorId": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "icd10",
          "nameUz",
          "synonyms",
          "category",
          "audienceLevels",
          "severityLevels",
          "protocolSources",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "DiseaseListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiseaseResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ]
      },
      "CreateDiseaseDto": {
        "type": "object",
        "properties": {
          "icd10": {
            "type": "string",
            "description": "ICD-10 kodi",
            "example": "I10"
          },
          "icd11": {
            "type": "string",
            "description": "ICD-11 kodi"
          },
          "nameUz": {
            "type": "string",
            "description": "Kasallik nomi (uz)",
            "example": "Gipertoniya"
          },
          "nameRu": {
            "type": "string",
            "description": "Kasallik nomi (ru)"
          },
          "nameEn": {
            "type": "string",
            "description": "Kasallik nomi (en)"
          },
          "nameLat": {
            "type": "string",
            "description": "Lotin tilidagi nomi"
          },
          "synonyms": {
            "description": "Sinonimlar",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "category": {
            "type": "string",
            "description": "Kategoriya",
            "example": "cardiology"
          },
          "audienceLevels": {
            "type": "array",
            "description": "Auditoriya darajalari",
            "items": {
              "type": "string",
              "enum": [
                "PATIENT",
                "STUDENT",
                "NURSE",
                "DOCTOR",
                "SPECIALIST",
                "MIXED"
              ]
            }
          },
          "severityLevels": {
            "type": "array",
            "description": "Kasallik og'irligi darajalari",
            "items": {
              "type": "string",
              "enum": [
                "MILD",
                "MODERATE",
                "SEVERE",
                "CRITICAL"
              ]
            }
          },
          "protocolSources": {
            "description": "Protokol manbalari",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "description": "Tasdiqlash statusi",
            "enum": [
              "DRAFT",
              "REVIEW",
              "APPROVED",
              "PUBLISHED",
              "REJECTED",
              "ARCHIVED"
            ]
          },
          "editorId": {
            "type": "number",
            "description": "Tahrirchi user ID (server o'rnatadi)"
          }
        },
        "required": [
          "icd10",
          "nameUz",
          "category"
        ]
      },
      "UpdateDiseaseDto": {
        "type": "object",
        "properties": {
          "icd10": {
            "type": "string",
            "description": "ICD-10 kodi"
          },
          "icd11": {
            "type": "string"
          },
          "nameUz": {
            "type": "string"
          },
          "nameRu": {
            "type": "string"
          },
          "nameEn": {
            "type": "string"
          },
          "nameLat": {
            "type": "string"
          },
          "synonyms": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "category": {
            "type": "string"
          },
          "audienceLevels": {
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "severityLevels": {
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "protocolSources": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "REVIEW",
              "APPROVED",
              "PUBLISHED",
              "REJECTED",
              "ARCHIVED"
            ]
          },
          "editorId": {
            "type": "number"
          }
        }
      },
...
      "CreateScientistDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Olim to'liq ismi",
            "example": "Scipione Riva-Rocci"
          },
          "role": {
            "type": "string",
            "description": "Roli",
            "enum": [
              "DISCOVERER",
              "CLASSIFIER",
              "CONTRIBUTOR",
              "RESEARCHER",
              "EDITOR"
            ],
            "example": "DISCOVERER"
          },
          "country": {
            "type": "string",
            "description": "Davlat",
            "example": "Italiya"
          },
          "birthYear": {
            "type": "number",
            "description": "Tug'ilgan yil",
            "example": 1863
          },
          "deathYear": {
            "type": "number",
            "description": "Vafot etgan yil",
            "example": 1937
          },
          "bioMd": {
            "type": "string",
            "description": "Biografiya (Markdown)"
          },
          "contributionsMd": {
            "type": "string",
            "description": "Hissasi (Markdown)"
          },
          "photoUrl": {
            "type": "string",
            "description": "Rasm URL"
          },
          "orderIndex": {
            "type": "number",
            "description": "Tartib indeksi",
            "default": 0
          }
        },
        "required": [
          "fullName",
          "role"
        ]
      },
      "UpdateScientistDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "description": "Olim to'liq ismi",
            "example": "Scipione Riva-Rocci"
          },
          "role": {
            "type": "string",
            "description": "Roli",
            "enum": [
              "DISCOVERER",
              "CLASSIFIER",
              "CONTRIBUTOR",
              "RESEARCHER",
              "EDITOR"
            ],
            "example": "DISCOVERER"
          },
          "country": {
            "type": "string",
            "description": "Davlat",
            "example": "Italiya"
          },
          "birthYear": {
            "type": "number",
            "description": "Tug'ilgan yil",
            "example": 1863
          },
          "deathYear": {
            "type": "number",
            "description": "Vafot etgan yil",
            "example": 1937
          },
          "bioMd": {
            "type": "string",
            "description": "Biografiya (Markdown)"
          },
          "contributionsMd": {
            "type": "string",
            "description": "Hissasi (Markdown)"
          },
          "photoUrl": {
            "type": "string",
            "description": "Rasm URL"
          },
          "orderIndex": {
            "type": "number",
            "description": "Tartib indeksi",
            "default": 0
          }
        }
      },
      "CreateResearchDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Tadqiqot sarlavhasi"
          },
          "authors": {
            "type": "string",
            "description": "Mualliflar",
            "example": "Smith J, Brown A, et al."
          },
          "journal": {
            "type": "string",
            "description": "Jurnal"
          },
          "year": {
            "type": "number",
            "description": "Nashr etilgan yil",
            "example": 2024
          },
          "doi": {
            "type": "string",
            "description": "DOI",
            "example": "10.1056/NEJMoa1511939"
          },
          "pubmedId": {
            "type": "string",
            "description": "PubMed ID"
          },
          "nctId": {
            "type": "string",
            "description": "ClinicalTrials.gov NCT ID"
          },
          "type": {
            "type": "string",
            "description": "Tadqiqot turi",
            "enum": [
              "RCT",
              "META_ANALYSIS",
              "SYSTEMATIC_REVIEW",
              "COHORT",
              "CASE_CONTROL",
              "CASE_SERIES",
              "CASE_REPORT",
              "GUIDELINE"
            ]
          },
          "summaryMd": {
            "type": "string",
            "description": "Qisqacha mazmun (Markdown)"
          },
          "evidenceLevel": {
            "type": "string",
            "description": "Dalil darajasi",
            "enum": [
              "A",
              "B",
              "C",
              "D"
            ],
            "default": "C"
          },
          "isLandmark": {
            "type": "boolean",
            "description": "Landmark tadqiqot",
            "default": false
          }
        },
        "required": [
          "title",
          "authors",
          "year",
          "type",
          "summaryMd"
        ]
      },
...
    }
  }
}